#240  Create page context. Move base contexts to page module side
