+++ pages/newcontext/__init__.py
@@ -1,0 +1,3 @@
+from . import base, pages
+from .base import Context, Contexts
+from .pages import Page
